pyproject.toml
[project]
name = "modle-service"
version = "0.1.0"
description = ""
authors = [
    {name = "sjc",email = "317897874@qq.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "faster-whisper (>=1.2.0,<2.0.0)",
    "uvicorn (>=0.32.1,<0.33.0)",
    "sqlalchemy (>=2.0.36,<3.0.0)",
    "pydantic-settings (>=2.6.1,<3.0.0)",
    "python-multipart (>=0.0.12,<0.1.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "langchain-postgres (>=0.0.15,<0.0.16)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "openai (>=1.107.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
